requirements.txt
asgiref==3.6.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-googledrive-storage==1.6.0
djangorestframework==3.14.0
google-api-core==2.10.2
google-api-python-client==2.86.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
Pillow==9.4.0
protobuf==4.22.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.30.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
